+++ BCA_Only_04302025.py
