+++ Normalize_RedAlkDigest_04302025.py
